ReleaseBuilder first version.
Made standalone by including all RCore and not using external assemblies.

This version will create a zip file from a config, supports finding directories; e.g.

The target specified on the command line indicates which Target to use in the XML. The target defines where to put the final archive and also optionally can run a transform to get the archive version which by default will use the gitversion.

Files and Folders will be located from the Root (if not in the current directory). When using a variable in an artefact with a wildcard a directory scan will be used to locate the files. Multiple artefacts can be included.

The list of files from the artefacts will have duplicates removed before creating the zip file.

<?xml version="1.0" encoding="utf-8"?>
<ReleaseConfig p1:noNamespaceSchemaLocation="ReleaseConfig.xsd" xmlns:p1="http://www.w3.org/2001/XMLSchema-instance">
    <Name>ThemisApp</Name>
    <Folder name="APPPATH" path="MSOS.UWP_*" version="latest" />
    <Target name="test" path="$SYNC_MSOS_TEST" archive-version="GetVersion,~APPPATH~" />
    <Target name="live" path="$SYNC_MSOS_LIVE" archive-version="GetVersion,~APPPATH~" />
    <Target name="demo" path="$SYNC_MSOS_DEMO" archive-version="GetVersion,~APPPATH~" />
    <Artefact file="index.html" />
    <Artefact file="MSOS.UWP.appinstaller" />
    <Artefact folder="~APPPATH~/*.appxbundle" />
    <Artefact folder="~APPPATH~/Dependencies/x64/*"/>
</ReleaseConfig>