ReleaseBuilder changes
New command line switches
--toolsdir specify where the tools are (default ..\release-tools)
--verbose twice will increase verbosity
--nobuild will skip the new build step

Reworked the config so that there is now an Artefacts tag that contains

<build> steps
<file> or folder artefacts to add

The build section can contain
<clean folder="files/"/>
<copy from="publish-web" [to="<folder>"] [name="newname.ext"]/>
<exec app="<tools file>" args="<app args>"/>

Each step will be performed in turn; so there can be extra <copy> after an exec to copy items created by the exec.

Changed the artefact so that it now remembers where it came from so this part of the path can be stripped when adding to the zipfile.