fixed related to angular release
Need to use ShellExecute (with correct path) to use node. For some reason ng still doesn't work.

Revised to permit multiple paths for tools; which in turn is added to the internal path to search for tools.

Fix bug getting folder contents where it would trim off the last directory because used GetDirectoryName and the method isn't clever enough to know that the last part is a directory...